+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - added XAUTHORITY environment
 - changed Nix runtime definition to contain a single installable
 - added support for local flake usage
-- added support for nix develop devshell
+- added support for nix develop devshell and nix shell
 - added flake program path discovery and caching
 
 ## 0.6.0 (2024-10-16)